refactor(components): consolidate date utility imports and remove unused file
- Moved getSingaporeDateYYYYMMDD and getSingaporeDateLong imports to a shared utils file.
- Removed the now redundant utils.ts file from DynamicDataBanner.
- Updated references in DynamicDataBanner and Homepage stories to reflect the new import paths.
