--- packages/components/src/templates/next/components/complex/DynamicDataBanner/utils.ts
@@ -1,4 +1,0 @@
-export {
-  getSingaporeDateLong,
-  getSingaporeDateYYYYMMDD,
-} from "~/utils/getSingaporeDate"
